requirements.txt
django==2.1.1
django-ckeditor==5.6.1
django-cleanup==2.1.0
django-stdimage==4.0.0
django-recaptcha==1.4.0
django-crispy-forms==1.7.2
Pillow==5.3.0
